Add AuctionService and move Extra Pets to it
- Fix the pet species short->int coercion which was making the database index not be used

+++ apps/web/Models/UserAuctionData.cs
@@ -8,9 +8,9 @@ public class UserAuctionData
 {
     public Dictionary<int, List<WowAuction>> RawAuctions { get; set; }
     public Dictionary<int, string> Names { get; set; }
-        
+
     [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-    public Dictionary<int, UserAuctionDataPet[]> Pets { get; set; }
+    public Dictionary<short, UserAuctionDataPet[]> Pets { get; set; }
 }
 
 public class UserAuctionDataPet
@@ -19,7 +19,7 @@ public class UserAuctionDataPet
     public int Level { get; set; }
     public ItemLocation Location { get; set; }
     public WowQuality Quality { get; set; }
-        
+
     [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
     public int? LocationId { get; set; }
